+++ tests/utils/test_caching.py
@@ -1,4 +1,6 @@
 import logging
+import pickle
+import tempfile
 from time import sleep, time
 from typing import Optional
 
@@ -7,164 +9,324 @@ import pytest
 from numpy.typing import NDArray
 
 from pydvl.parallel import MapReduceJob
-from pydvl.utils import memcached
+from pydvl.utils.caching import (
+    CachedFunc,
+    CachedFuncConfig,
+    DiskCacheBackend,
+    InMemoryCacheBackend,
+    MemcachedCacheBackend,
+)
 from pydvl.utils.types import Seed
 
 logger = logging.getLogger(__name__)
 
 
-def test_failed_connection():
-    from pydvl.utils import MemcachedClientConfig
-
-    client_config = MemcachedClientConfig(server=("localhost", 0), connect_timeout=0.1)
-    with pytest.raises((ConnectionRefusedError, OSError)):
-        memcached(client_config)(lambda x: x)
+def foo(indices: NDArray[np.int_], *args, **kwargs) -> float:
+    return float(np.sum(indices))
 
 
-def test_memcached_single_job(memcached_client):
-    client, config = memcached_client
+def foo_duplicate(indices: NDArray[np.int_], *args, **kwargs) -> float:
+    return float(np.sum(indices))
 
-    # TODO: maybe this should be a fixture too...
-    @memcached(client_config=config, time_threshold=0)  # Always cache results
-    def foo(indices: NDArray[np.int_]) -> float:
-        return float(np.sum(indices))
+
+def foo_with_random(indices: NDArray[np.int_], *args, **kwargs) -> float:
+    rng = np.random.default_rng()
+    scale = kwargs.get("scale", 1.0)
+    return float(np.sum(indices)) + rng.normal(scale=scale)
+
+
+def foo_with_random_and_sleep(indices: NDArray[np.int_], *args, **kwargs) -> float:
+    sleep(0.01)
+    rng = np.random.default_rng()
+    scale = kwargs.get("scale", 1.0)
+    return float(np.sum(indices)) + rng.normal(scale=scale)
+
+
+# Used to test caching of methods
+class CacheTest:
+    def __init__(self):
+        self.value = 0
+
+    def foo(self):
+        return 1
+
+
+@pytest.fixture(params=["in-memory", "disk", "memcached"])
+def cache_backend(request):
+    backend: str = request.param
+    if backend == "in-memory":
+        cache_backend = InMemoryCacheBackend()
+        yield cache_backend
+        cache_backend.clear()
+    elif backend == "disk":
+        with tempfile.TemporaryDirectory() as tempdir:
+            cache_backend = DiskCacheBackend(tempdir)
+            yield cache_backend
+            cache_backend.clear()
+    elif backend == "memcached":
+        cache_backend = MemcachedCacheBackend()
+        yield cache_backend
+        cache_backend.clear()
+    else:
+        raise ValueError(f"Unknown cache backend {backend}")
+
+
+@pytest.mark.parametrize(
+    "f1, f2, expected_equal",
+    [
+        # Test that the same function gets the same hash
+        (lambda x: x, lambda x: x, True),
+        (foo, foo, True),
+        # Test that different functions get different hashes
+        (foo, lambda x: x, False),
+        (foo, foo_with_random, False),
+        (foo_with_random, foo_with_random_and_sleep, False),
+        # Test that functions with different names but the same body get different hashes
+        (foo, foo_duplicate, True),
+    ],
+)
+def test_cached_func_hash_function(f1, f2, expected_equal):
+    f1_hash = CachedFunc._hash_function(f1)
+    f2_hash = CachedFunc._hash_function(f2)
+    if expected_equal:
+        assert f1_hash == f2_hash, f"{f1_hash} != {f2_hash}"
+    else:
+        assert f1_hash != f2_hash, f"{f1_hash} == {f2_hash}"
+
+
+@pytest.mark.parametrize(
+    "args1, args2, expected_equal",
+    [
+        # Test that the same arguments get the same hash
+        ([[]], [[]], True),
+        ([[1]], [[1]], True),
+        ([frozenset([])], [frozenset([])], True),
+        ([frozenset([1])], [frozenset([1])], True),
+        ([np.ones(3)], [np.ones(3)], True),
+        ([np.ones(3), 16], [np.ones(3), 16], True),
+        ([frozenset(np.ones(3))], [frozenset(np.ones(3))], True),
+        # Test that different arguments get different hashes
+        ([[1, 2, 3]], [np.ones(3)], False),
+        ([np.ones(3)], [np.ones(5)], False),
+        ([np.ones(3)], [frozenset(np.ones(3))], False),
+    ],
+)
+def test_cached_func_hash_arguments(args1, args2, expected_equal):
+    args1_hash = CachedFunc._hash_arguments(foo, ignore_args=[], args=args1, kwargs={})
+    args2_hash = CachedFunc._hash_arguments(foo, ignore_args=[], args=args2, kwargs={})
+    if expected_equal:
+        assert args1_hash == args2_hash, f"{args1_hash} != {args2_hash}"
+    else:
+        assert args1_hash != args2_hash, f"{args1_hash} == {args2_hash}"
+
+
+def test_cached_func_hash_arguments_of_method():
+    obj = CacheTest()
+
+    hash1 = CachedFunc._hash_arguments(obj.foo, [], tuple(), {})
+    obj.value += 1
+    hash2 = CachedFunc._hash_arguments(obj.foo, [], tuple(), {})
+    assert hash1 == hash2
+
+
+def test_cache_backend_serialization(cache_backend):
+    value = 16.8
+    cache_backend.set("key", value)
+    deserialized_cache_backend = pickle.loads(pickle.dumps(cache_backend))
+    assert deserialized_cache_backend.get("key") == value
+    if isinstance(cache_backend, InMemoryCacheBackend):
+        assert cache_backend.cached_values == deserialized_cache_backend.cached_values
+    elif isinstance(cache_backend, DiskCacheBackend):
+        assert cache_backend.cache_dir == deserialized_cache_backend.cache_dir
+
+
+def test_single_job(cache_backend):
+    cached_func_config = CachedFuncConfig(time_threshold=0.0)
+    wrapped_foo = cache_backend.wrap(foo, config=cached_func_config)
 
     n = 1000
-    foo(np.arange(n))
-    hits_before = client.stats()[b"get_hits"]
-    foo(np.arange(n))
-    hits_after = client.stats()[b"get_hits"]
+    wrapped_foo(np.arange(n))
+    hits_before = wrapped_foo.stats.hits
+    wrapped_foo(np.arange(n))
+    hits_after = wrapped_foo.stats.hits
 
     assert hits_after > hits_before
 
 
-def test_memcached_parallel_jobs(memcached_client, parallel_config):
-    client, config = memcached_client
+def test_without_pymemcache(mocker):
+    mocker.patch("pydvl.utils.caching.memcached.PYMEMCACHE_INSTALLED", False)
+    with pytest.raises(ModuleNotFoundError):
+        MemcachedCacheBackend()
 
-    @memcached(
-        client_config=config,
-        time_threshold=0,  # Always cache results
-        # Note that we typically do NOT want to ignore run_id
+
+def test_memcached_failed_connection():
+    from pydvl.utils import MemcachedClientConfig
+
+    config = MemcachedClientConfig(server=("localhost", 0), connect_timeout=0.1)
+    with pytest.raises((ConnectionRefusedError, OSError)):
+        MemcachedCacheBackend(config)
+
+
+def test_cache_time_threshold(cache_backend):
+    cached_func_config = CachedFuncConfig(time_threshold=1.0)
+    wrapped_foo = cache_backend.wrap(foo, config=cached_func_config)
+
+    n = 1000
+    indices = np.arange(n)
+    wrapped_foo(indices)
+    hits_before = wrapped_foo.stats.hits
+    misses_before = wrapped_foo.stats.misses
+    wrapped_foo(indices)
+    hits_after = wrapped_foo.stats.hits
+    misses_after = wrapped_foo.stats.misses
+
+    assert hits_after == hits_before
+    assert misses_after > misses_before
+
+
+def test_cache_ignore_args(cache_backend):
+    # Note that we typically do NOT want to ignore run_id
+    cached_func_config = CachedFuncConfig(
+        time_threshold=0.0,
+        ignore_args=["job_id"],
+    )
+    wrapped_foo = cache_backend.wrap(foo, config=cached_func_config)
+
+    n = 1000
+    indices = np.arange(n)
+    wrapped_foo(indices, job_id=0)
+    hits_before = wrapped_foo.stats.hits
+    wrapped_foo(indices, job_id=16)
+    hits_after = wrapped_foo.stats.hits
+
+    assert hits_after > hits_before
+
+
+def test_parallel_jobs(cache_backend, parallel_config):
+    if not isinstance(cache_backend, MemcachedCacheBackend):
+        pytest.skip("Only running this test with MemcachedCacheBackend")
+    if parallel_config.backend != "joblib":
+        pytest.skip("We don't have to test this with all parallel backends")
+
+    # Note that we typically do NOT want to ignore run_id
+    cached_func_config = CachedFuncConfig(
         ignore_args=["job_id", "run_id"],
     )
-    def foo(indices: NDArray[np.int_], *args, **kwargs) -> float:
-        # logger.info(f"run_id: {run_id}, running...")
-        return float(np.sum(indices))
+    wrapped_foo = cache_backend.wrap(foo, config=cached_func_config)
 
     n = 1234
     n_runs = 10
-    hits_before = client.stats()[b"get_hits"]
+    hits_before = cache_backend.client.stats()[b"get_hits"]
 
     map_reduce_job = MapReduceJob(
-        np.arange(n), foo, np.sum, n_jobs=4, config=parallel_config
+        np.arange(n), wrapped_foo, np.sum, n_jobs=4, config=parallel_config
     )
     results = []
 
     for _ in range(n_runs):
         result = map_reduce_job()
         results.append(result)
-    hits_after = client.stats()[b"get_hits"]
+    hits_after = cache_backend.client.stats()[b"get_hits"]
 
     assert results[0] == n * (n - 1) / 2  # Sanity check
     # FIXME! This is non-deterministic: if packets are delayed for longer than
     #  the timeout configured then we won't have num_runs hits. So we add this
     #  good old hard-coded magic number here.
-    assert hits_after - hits_before >= n_runs - 2
+    assert hits_after - hits_before >= n_runs - 2, wrapped_foo.stats
 
 
-def test_memcached_repeated_training(memcached_client, worker_id: str):
-    _, config = memcached_client
-
-    @memcached(
-        client_config=config,
-        time_threshold=0,  # Always cache results
+def test_repeated_training(cache_backend, worker_id: str):
+    cached_func_config = CachedFuncConfig(
+        time_threshold=0.0,
         allow_repeated_evaluations=True,
         rtol_stderr=0.01,
     )
-    def foo(indices: NDArray[np.int_], uid: str) -> float:
-        return float(np.sum(indices)) + np.random.normal(scale=10)
+    wrapped_foo = cache_backend.wrap(
+        foo_with_random,
+        config=cached_func_config,
+    )
 
     n = 7
-    foo(np.arange(n), worker_id)
-    for _ in range(10_000):
-        result = foo(np.arange(n), worker_id)
+    indices = np.arange(n)
 
-    assert (result - np.sum(np.arange(n))) < 1
-    assert foo.stats.sets < foo.stats.hits
+    for _ in range(1_000):
+        result = wrapped_foo(indices, worker_id)
+
+    assert np.isclose(result, np.sum(indices), atol=1)
+    assert wrapped_foo.stats.sets < wrapped_foo.stats.hits
 
 
-def test_memcached_faster_with_repeated_training(memcached_client, worker_id: str):
-    _, config = memcached_client
-
-    @memcached(
-        client_config=config,
-        time_threshold=0,  # Always cache results
+def test_faster_with_repeated_training(cache_backend, worker_id: str):
+    cached_func_config = CachedFuncConfig(
+        time_threshold=0.0,
         allow_repeated_evaluations=True,
         rtol_stderr=0.1,
     )
-    def foo_cache(indices: NDArray[np.int_], uid: str) -> float:
-        sleep(0.01)
-        return float(np.sum(indices)) + np.random.normal(scale=1)
-
-    def foo_no_cache(indices: NDArray[np.int_], uid: str) -> float:
-        sleep(0.01)
-        return float(np.sum(indices)) + np.random.normal(scale=1)
+    wrapped_foo = cache_backend.wrap(
+        foo_with_random_and_sleep,
+        config=cached_func_config,
+    )
 
     n = 3
-    foo_cache(np.arange(n), worker_id)
-    foo_no_cache(np.arange(n), worker_id)
+    n_repetitions = 500
+    indices = np.arange(n)
 
     start = time()
-    for _ in range(300):
-        result_fast = foo_cache(np.arange(n), worker_id)
+    for _ in range(n_repetitions):
+        result_fast = wrapped_foo(indices, worker_id)
     end = time()
     fast_time = end - start
 
     start = time()
     results_slow = []
-    for _ in range(300):
-        result = foo_no_cache(np.arange(n), worker_id)
+    for _ in range(n_repetitions):
+        result = foo_with_random_and_sleep(indices, worker_id)
         results_slow.append(result)
     end = time()
     slow_time = end - start
 
-    assert (result_fast - np.mean(results_slow)) < 1
+    assert np.isclose(np.mean(results_slow), np.sum(indices), atol=0.1)
+    assert np.isclose(result_fast, np.mean(results_slow), atol=1)
     assert fast_time < slow_time
 
 
 @pytest.mark.parametrize("n, atol", [(10, 5), (20, 10)])
 @pytest.mark.parametrize("n_jobs", [1, 2])
 @pytest.mark.parametrize("n_runs", [20])
-def test_memcached_parallel_repeated_training(
-    memcached_client, n, atol, n_jobs, n_runs, parallel_config, seed
+def test_parallel_repeated_training(
+    cache_backend, n, atol, n_jobs, n_runs, parallel_config
 ):
     if parallel_config.backend != "joblib":
         pytest.skip("We don't have to test this with all parallel backends")
-    _, config = memcached_client
 
-    @memcached(
-        client_config=config,
-        time_threshold=0,  # Always cache results
+    def map_func(indices: NDArray[np.int_], seed: Optional[Seed] = None) -> float:
+        return np.sum(indices).item() + np.random.normal(scale=1)
+
+    # Note that we typically do NOT want to ignore run_id
+    cached_func_config = CachedFuncConfig(
+        time_threshold=0.0,
         allow_repeated_evaluations=True,
         rtol_stderr=0.01,
-        # Note that we typically do NOT want to ignore run_id
         ignore_args=["job_id", "run_id"],
     )
-    def map_func(indices: NDArray[np.int_], seed: Optional[Seed] = None) -> float:
-        # from pydvl.utils.logging import logger
-        # logger.info(f"run_id: {run_id}, running...")
-        rng = np.random.default_rng(seed)
-        return np.sum(indices).item() + rng.normal(scale=5)
+    wrapped_map_func = cache_backend.wrap(
+        map_func,
+        config=cached_func_config,
+    )
 
     def reduce_func(chunks: NDArray[np.float_]) -> float:
         return np.sum(chunks).item()
 
     map_reduce_job = MapReduceJob(
-        np.arange(n), map_func, reduce_func, n_jobs=n_jobs, config=parallel_config
+        np.arange(n),
+        wrapped_map_func,
+        reduce_func,
+        n_jobs=n_jobs,
+        config=parallel_config,
     )
     results = []
     for _ in range(n_runs):
-        result = map_reduce_job(seed=seed)
+        result = map_reduce_job()
         results.append(result)
 
     exact_value = np.sum(np.arange(n)).item()